backend2/requirements.txt
flask
flask_cors
python-dotenv
requests